Move population table control component to own file
